Refactored Maven generation mojos to more effeciently handle refreshing the IDE (through m2e) for resources that changed. Also ensured that the stale file is created, even when a processing error occurs. This may fix some runaway problems in Eclipse.
