+++ metaschema-maven-plugin/src/main/java/gov/nist/secauto/metaschema/maven/plugin/GenerateSchemaMojo.java
@@ -9,9 +9,8 @@ import gov.nist.secauto.metaschema.core.configuration.DefaultConfiguration;
 import gov.nist.secauto.metaschema.core.configuration.IConfiguration;
 import gov.nist.secauto.metaschema.core.configuration.IMutableConfiguration;
 import gov.nist.secauto.metaschema.core.model.IModule;
-import gov.nist.secauto.metaschema.core.model.MetaschemaException;
+import gov.nist.secauto.metaschema.core.util.CollectionUtil;
 import gov.nist.secauto.metaschema.core.util.ObjectUtils;
-import gov.nist.secauto.metaschema.databind.IBindingContext;
 import gov.nist.secauto.metaschema.schemagen.ISchemaGenerator;
 import gov.nist.secauto.metaschema.schemagen.SchemaGenerationFeature;
 import gov.nist.secauto.metaschema.schemagen.json.JsonSchemaGenerator;
@@ -30,6 +29,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardOpenOption;
 import java.util.EnumSet;
+import java.util.LinkedList;
 import java.util.List;
 import java.util.Locale;
 import java.util.Set;
@@ -131,7 +131,9 @@ public class GenerateSchemaMojo
    * @throws MojoExecutionException
    *           if an error occurred during generation
    */
-  protected void generate(@NonNull Set<IModule> modules) throws MojoExecutionException {
+  @Override
+  @NonNull
+  protected List<File> generate(@NonNull Set<IModule> modules) throws MojoExecutionException {
     IMutableConfiguration<SchemaGenerationFeature<?>> schemaGenerationConfig
         = new DefaultConfiguration<>();
 
@@ -167,6 +169,7 @@ public class GenerateSchemaMojo
     }
 
     Path outputDirectory = ObjectUtils.notNull(getOutputDirectory().toPath());
+    List<File> generatedSchemas = new LinkedList<>();
     for (IModule module : modules) {
       if (getLog().isInfoEnabled()) {
         getLog().info(String.format("Processing metaschema: %s", module.getLocation()));
@@ -174,12 +177,14 @@ public class GenerateSchemaMojo
       if (module.getExportedRootAssemblyDefinitions().isEmpty()) {
         continue;
       }
-      generateSchemas(module, schemaGenerationConfig, outputDirectory, schemaFormats);
+      generatedSchemas.addAll(generateSchemas(module, schemaGenerationConfig, outputDirectory, schemaFormats));
     }
+    return CollectionUtil.unmodifiableList(generatedSchemas);
   }
 
   @SuppressWarnings("PMD.AvoidCatchingGenericException")
-  private static void generateSchemas(
+  @NonNull
+  private List<File> generateSchemas(
       @NonNull IModule module,
       @NonNull IConfiguration<SchemaGenerationFeature<?>> schemaGenerationConfig,
       @NonNull Path outputDirectory,
@@ -187,11 +192,13 @@ public class GenerateSchemaMojo
 
     String shortName = module.getShortName();
 
+    List<File> generatedSchemas = new LinkedList<>();
     if (schemaFormats.contains(SchemaFormat.XSD)) {
       try { // XML Schema
         String filename = String.format("%s_schema.xsd", shortName);
         Path xmlSchema = ObjectUtils.notNull(outputDirectory.resolve(filename));
         generateSchema(module, schemaGenerationConfig, xmlSchema, XML_SCHEMA_GENERATOR);
+        generatedSchemas.add(xmlSchema.toFile());
       } catch (Exception ex) {
         throw new MojoExecutionException("Unable to generate XML schema.", ex);
       }
@@ -200,12 +207,14 @@ public class GenerateSchemaMojo
     if (schemaFormats.contains(SchemaFormat.JSON_SCHEMA)) {
       try { // JSON Schema
         String filename = String.format("%s_schema.json", shortName);
-        Path xmlSchema = ObjectUtils.notNull(outputDirectory.resolve(filename));
-        generateSchema(module, schemaGenerationConfig, xmlSchema, JSON_SCHEMA_GENERATOR);
+        Path jsonSchema = ObjectUtils.notNull(outputDirectory.resolve(filename));
+        generateSchema(module, schemaGenerationConfig, jsonSchema, JSON_SCHEMA_GENERATOR);
+        generatedSchemas.add(jsonSchema.toFile());
       } catch (Exception ex) {
         throw new MojoExecutionException("Unable to generate JSON schema.", ex);
       }
     }
+    return CollectionUtil.unmodifiableList(generatedSchemas);
   }
 
   private static void generateSchema(
@@ -222,76 +231,5 @@ public class GenerateSchemaMojo
         StandardOpenOption.TRUNCATE_EXISTING))) {
       generator.generateFromModule(module, writer, schemaGenerationConfig);
     }
-  }
-
-  @Override
-  public void execute() throws MojoExecutionException {
-    File staleFile = getStaleFile();
-    try {
-      staleFile = ObjectUtils.notNull(staleFile.getCanonicalFile());
-    } catch (IOException ex) {
-      if (getLog().isWarnEnabled()) {
-        getLog().warn("Unable to resolve canonical path to stale file. Treating it as not existing.", ex);
-      }
-    }
-
-    boolean generate;
-    if (shouldExecutionBeSkipped()) {
-      if (getLog().isDebugEnabled()) {
-        getLog().debug(String.format("Schema generation is configured to be skipped. Skipping."));
-      }
-      generate = false;
-    } else if (staleFile.exists()) {
-      generate = isGenerationRequired();
-    } else {
-      if (getLog().isInfoEnabled()) {
-        getLog().info(String.format("Stale file '%s' doesn't exist! Generating source files.", staleFile.getPath()));
-      }
-      generate = true;
-    }
-
-    if (generate) {
-      performGeneration();
-      createStaleFile(staleFile);
-
-      // for m2e
-      getBuildContext().refresh(getOutputDirectory());
-    }
-    // // add generated sources to Maven
-    // try {
-    // getMavenProject()..addCompileSourceRoot(getOutputDirectory().getCanonicalFile().getPath());
-    // } catch (IOException ex) {
-    // throw new MojoExecutionException("Unable to add output directory to maven
-    // sources.", ex);
-    // }
-  }
-
-  @SuppressWarnings("PMD.AvoidCatchingGenericException")
-  private void performGeneration() throws MojoExecutionException {
-    File outputDir = getOutputDirectory();
-    if (getLog().isDebugEnabled()) {
-      getLog().debug(String.format("Using outputDirectory: %s", outputDir.getPath()));
-    }
-
-    if (!outputDir.exists() && !outputDir.mkdirs()) {
-      throw new MojoExecutionException("Unable to create output directory: " + outputDir);
-    }
-
-    IBindingContext bindingContext;
-    try {
-      bindingContext = newBindingContext();
-    } catch (MetaschemaException | IOException ex) {
-      throw new MojoExecutionException("Failed to create the binding context", ex);
-    }
-
-    // generate schemas based on provided metaschema sources
-    Set<IModule> modules;
-    try {
-      modules = getModulesToGenerateFor(bindingContext);
-    } catch (Exception ex) {
-      throw new MojoExecutionException("Loading of metaschema modules failed", ex);
-    }
-
-    generate(modules);
   }
 }
