+++ metaschema-maven-plugin/src/main/java/gov/nist/secauto/metaschema/maven/plugin/GenerateSourcesMojo.java
@@ -6,9 +6,8 @@
 package gov.nist.secauto.metaschema.maven.plugin;
 
 import gov.nist.secauto.metaschema.core.model.IModule;
-import gov.nist.secauto.metaschema.core.model.MetaschemaException;
 import gov.nist.secauto.metaschema.core.util.ObjectUtils;
-import gov.nist.secauto.metaschema.databind.IBindingContext;
+import gov.nist.secauto.metaschema.databind.codegen.IProduction;
 import gov.nist.secauto.metaschema.databind.codegen.JavaGenerator;
 import gov.nist.secauto.metaschema.databind.codegen.config.DefaultBindingConfiguration;
 
@@ -23,6 +22,7 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
 
@@ -78,7 +78,9 @@ public class GenerateSourcesMojo
    * @throws MojoExecutionException
    *           if an error occurred while generating sources
    */
-  protected void generate(@NonNull Set<IModule> modules) throws MojoExecutionException {
+  @Override
+  @NonNull
+  protected List<File> generate(@NonNull Set<IModule> modules) throws MojoExecutionException {
     DefaultBindingConfiguration bindingConfiguration = new DefaultBindingConfiguration();
     for (File config : getConfigs()) {
       try {
@@ -92,50 +94,17 @@ public class GenerateSourcesMojo
       }
     }
 
+    IProduction production;
     try {
       if (getLog().isInfoEnabled()) {
         getLog().info("Generating Java classes in: " + getOutputDirectory().getPath());
       }
-      JavaGenerator.generate(modules, ObjectUtils.notNull(getOutputDirectory().toPath()),
+      production = JavaGenerator.generate(
+          modules,
+          ObjectUtils.notNull(getOutputDirectory().toPath()),
           bindingConfiguration);
     } catch (IOException ex) {
       throw new MojoExecutionException("Creation of Java classes failed.", ex);
-    }
-  }
-
-  @SuppressWarnings("PMD.AvoidCatchingGenericException")
-  @Override
-  public void execute() throws MojoExecutionException {
-    File staleFile = getStaleFile();
-    try {
-      staleFile = ObjectUtils.notNull(staleFile.getCanonicalFile());
-    } catch (IOException ex) {
-      if (getLog().isWarnEnabled()) {
-        getLog().warn("Unable to resolve canonical path to stale file. Treating it as not existing.", ex);
-      }
-    }
-
-    boolean generate;
-    if (shouldExecutionBeSkipped()) {
-      if (getLog().isDebugEnabled()) {
-        getLog().debug(String.format("Source file generation is configured to be skipped. Skipping."));
-      }
-      generate = false;
-    } else if (staleFile.exists()) {
-      generate = isGenerationRequired();
-    } else {
-      if (getLog().isInfoEnabled()) {
-        getLog().info(String.format("Stale file '%s' doesn't exist! Generating source files.", staleFile.getPath()));
-      }
-      generate = true;
-    }
-
-    if (generate) {
-      performGeneration();
-      createStaleFile(staleFile);
-
-      // for m2e
-      getBuildContext().refresh(getOutputDirectory());
     }
 
     // add generated sources to Maven
@@ -144,35 +113,9 @@ public class GenerateSourcesMojo
     } catch (IOException ex) {
       throw new MojoExecutionException("Unable to add output directory to maven sources.", ex);
     }
+
+    return ObjectUtils.notNull(production.getGeneratedClasses()
+        .map(gen -> gen.getClassFile().toFile())
+        .collect(Collectors.toUnmodifiableList()));
   }
-
-  @SuppressWarnings({ "PMD.AvoidCatchingGenericException", "PMD.ExceptionAsFlowControl" })
-  private void performGeneration() throws MojoExecutionException {
-    File outputDir = getOutputDirectory();
-    if (getLog().isDebugEnabled()) {
-      getLog().debug(String.format("Using outputDirectory: %s", outputDir.getPath()));
-    }
-
-    if (!outputDir.exists() && !outputDir.mkdirs()) {
-      throw new MojoExecutionException("Unable to create output directory: " + outputDir);
-    }
-
-    IBindingContext bindingContext;
-    try {
-      bindingContext = newBindingContext();
-    } catch (MetaschemaException | IOException ex) {
-      throw new MojoExecutionException("Failed to create the binding context", ex);
-    }
-
-    // generate Java sources based on provided metaschema sources
-    Set<IModule> modules;
-    try {
-      modules = getModulesToGenerateFor(bindingContext);
-    } catch (Exception ex) {
-      throw new MojoExecutionException("Loading of metaschema modules failed", ex);
-    }
-
-    generate(modules);
-  }
-
 }
